solve problem with non persisting default columns when creating boards. Modify file names, adding models at the beggining. Add setSingleColumn to routes.
