+++ api/boards/boards.services.js
@@ -1,4 +1,4 @@
-const Board = require('./model');
+const Board = require('./boards.model');
 
 function getAllBoard() {
   return Board.find({});
@@ -13,7 +13,7 @@ function createBoard(board) {
 }
 
 function updateBoard(id, board) {
-  return Board.findByIdAndUpdate(id, board);
+  return Board.findByIdAndUpdate(id, board, { new: true });
 }
 
 function deleteBoard(id) {
