--- api/cards/routes.js
@@ -1,22 +1,0 @@
-const express = require('express');
-const { isAuthenticated } = require('../middlewares/auth/services');
-
-const controller = require('./controller');
-
-const {
-  createCardHandler,
-  getAllCardHandler,
-  getSingleCardHandler,
-  updateCardHandler,
-  deleteCardHandler,
-} = controller;
-
-const router = express.Router();
-
-router.get('/', getAllCardHandler);
-router.post('/', isAuthenticated, createCardHandler);
-router.get('/:id', getSingleCardHandler);
-router.patch('/:id', isAuthenticated, updateCardHandler);
-router.delete('/:id', isAuthenticated, deleteCardHandler);
-
-module.exports = router;
